+++ src/PlaneEngine.h
@@ -6,7 +6,7 @@ class PlaneEngine : public Engine {
     public:
         PlaneEngine();
 
-        void move(int horDir, int vertDir);
+        void move(double horDir, double vertDir);
         Vector mapToGlobal(const Vector& pt) const;
         Vector mapToLocal(const Vector& pt) const;
         EngineType getType() const { return PlaneEngineType; }
