+++ src/SphereEngine.h
@@ -10,7 +10,7 @@ class SphereEngine : public Engine {
     public:
         SphereEngine();
 
-        void move(int horDir, int vertDir);
+        void move(double horDir, double vertDir);
         Vector mapToGlobal(const Vector& pt) const;
         Vector mapToLocal(const Vector& pt) const;
         EngineType getType() const { return SphereEngineType; }
